refactor mainactivity
